refactor(retrieval): own Sentence Transformers export metadata
